Rewrite GlobalNavElements in ts
